Enforce parameter names to match Puppet's naming requirement.
This allows the AST parameter class to throw a parse_exception if the
variable name does not conform to the same names as required by Puppet.
